180 flip update
Flipped this sucka and separated out the script, made a ton of updates
to that (too many to describe ATM). Also updated the look and feel of
the demo. Next up I'll actually be working out that crappy feedback
inline in the script and get some hot methods all up in this business.